move event, index, pom, config into trace
