+++ sk-core/src/lib.rs
@@ -1,17 +1,14 @@
 #![cfg_attr(coverage, feature(coverage_attribute))]
-pub mod config;
+
 pub mod constants;
 pub mod errors;
-pub mod event;
 pub mod events;
 pub mod external_storage;
 pub mod hooks;
-pub mod index;
 pub mod jsonutils;
 pub mod k8s;
 pub mod logging;
 pub mod macros;
-pub mod pod_owners_map;
 pub mod time;
 pub mod trace;
 
@@ -31,25 +28,25 @@ pub mod prelude {
         SimulationRoot,
     };
 
-    pub use crate::config::{
-        TracerConfig,
-        TrackedObjectConfig,
-    };
     pub use crate::constants::*;
     pub use crate::errors::EmptyResult;
-    pub use crate::event::{
-        TraceAction,
-        TraceEvent,
-        append_event,
-    };
     pub use crate::events::SkEventRecorder;
     pub use crate::k8s::{
         KubeResourceExt,
         OpenApiResourceExt,
     };
-    pub use crate::pod_owners_map::{
+    pub use crate::trace::Trace;
+    pub use crate::trace::config::{
+        TracerConfig,
+        TrackedObjectConfig,
+    };
+    pub use crate::trace::event::{
+        TraceAction,
+        TraceEvent,
+        append_event,
+    };
+    pub use crate::trace::pod_owners_map::{
         PodLifecyclesMap,
         PodOwnersMap,
     };
-    pub use crate::trace::Trace;
 }
