+++ sk-core/src/trace/tests/mod.rs
@@ -1,7 +1,4 @@
 mod exported_trace_test;
 mod pod_owners_map_test;
 
-use rstest::{
-    fixture,
-    rstest,
-};
+use sk_testutils::*;
