+++ sk-core/src/trace/tests/pod_owners_map_test.rs
@@ -3,12 +3,12 @@ use std::collections::HashMap;
 use super::*;
 use crate::constants::DEPLOYMENT_GVK;
 use crate::k8s::PodLifecycleData;
-use crate::pod_owners_map::{
+use crate::trace::TraceIndex;
+use crate::trace::pod_owners_map::{
     PodLifecyclesMap,
     PodOwnersMap,
     filter_lifecycles_map,
 };
-use crate::trace::TraceIndex;
 
 const START_TS: i64 = 5;
 const END_TS: i64 = 10;
