+++ sk-tracer/src/store.rs
@@ -3,7 +3,6 @@ use std::sync::Arc;
 
 use kube::Resource;
 use sk_api::v1::ExportFilters;
-use sk_core::index::TraceIndex;
 use sk_core::jsonutils;
 use sk_core::k8s::{
     DynamicApiSet,
@@ -15,6 +14,7 @@ use sk_core::k8s::{
     format_gvk_name,
 };
 use sk_core::prelude::*;
+use sk_core::trace::index::TraceIndex;
 use tokio::sync::Mutex;
 use tracing::*;
